requirements.txt
arrow==1.2.3
asgiref==3.5.2
blessed==1.19.1
certifi==2022.12.7
charset-normalizer==2.0.12
Django==3.2.13
django-cors-headers==3.7.0
django-filter==21.1
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.12.4
idna==3.4
Pillow==9.1.0
psycopg2==2.8.6
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
requests==2.27.1
schedule==1.1.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
uWSGI==2.0.20
wcwidth==0.2.5
